# foodapp/static/swagger.yaml
openapi: 3.0.1
info:
  title: FastFoodFastAPI
  description: Online Food Ordering Restaurant API
  version: '0.1'
servers:
  - url: http://127.0.0.1:5000
paths:
  /auth/login:

    post:
      description: Authenticates a registered user 
      consumes:
      - multipart/form-data
      produces:
      - application/json
      parameters:
      - in: formData
        name: email,
        description: Email,
        type: string
      - in: formData
        name: password,
        description: Password,
        type: string  
      tags:
        - Authentication
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                user_password:
                  type: string
                email:
                  type: string
            examples:
              '0':
                value: |2-
                   {
                             
                          "email":"nabatanzigorret143@gmail.com",
                          "user_password":"#golden@"
                   }
      responses:
        '200':
          description: Logged in successfully
        '404':
          description: Email not found
        '402':
          description: wrong password
        '201':
          description: Account disable
        '500':
          description: Internal Server error  
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  refresh_token:
                    type: string
                  user_email:
                    type: string
                  message:
                    type: string
              examples:
                '0':
                  value: |
                    {
                      "access_token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJmcmVzaCI6ZmFsc2UsImlhdCI6MTY0NzkzNTk1MSwianRpIjoiMGJjNjk1NzQtNjhlMS00MTg1LWE5ODktYjVhNGNjYjViNGIyIiwidHlwZSI6ImFjY2VzcyIsInN1YiI6NCwibmJmIjoxNjQ3OTM1OTUxLCJleHAiOjE2NDc5MzY4NTF9.ir-M4ncw1kuhgarQYMotKg6nf7qog1W4fgI4TZ3dpG8", 
                      "message": "You logged in successfully!", 
                      "refresh_token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJmcmVzaCI6ZmFsc2UsImlhdCI6MTY0NzkzNTk1MSwianRpIjoiNWEzM2FhZWYtNzY4NC00ZDk0LTkzNDctZGIwMTkyNzYxNTA0IiwidHlwZSI6InJlZnJlc2giLCJzdWIiOjQsIm5iZiI6MTY0NzkzNTk1MSwiZXhwIjoxNjUwNTI3OTUxfQ.IZ1-qZ5RIsySOIUi7xFn-V17jSFm-1jIRhvpF_9A3NI", 
                      "user_email": "nabatanzigorret143@gmail.com"
                    }
      servers:
        - url: http://127.0.0.1:5000
    servers:
      - url: http://127.0.0.1:5000
  /auth/token/refresh:
    post:
      description: Provides a new refresh token
      tags:
        - Authentication
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                user_password:
                  type: string
                email:
                  type: string
            examples:
              '0':
                value: |2-
                   {
                             
                          "email":"nabatanzigorret143@gmail.com",
                          "user_password":"#golden@"
                   }
      responses:
        '200':
          description: Auto generated using Swagger Inspector
          content:
            application/json:
              schema:
                type: object
                properties:
                  access:
                    type: string
              examples:
                '0':
                  value: |
                    {
                      "access": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJmcmVzaCI6ZmFsc2UsImlhdCI6MTY0NzkzNTk5OCwianRpIjoiM2Y4ZDZkZjctNDM1ZC00ZWNiLWFiZjYtOTY1ZWY4NjVhNmQzIiwidHlwZSI6ImFjY2VzcyIsInN1YiI6NCwibmJmIjoxNjQ3OTM1OTk4LCJleHAiOjE2NDc5MzY4OTh9.DoqKKqRiYUz_Ghy_OABDdDQITuFyNAfGmi3nA33Ardw"
                    }
      servers:
        - url: http://127.0.0.1:5000
    servers:
      - url: http://127.0.0.1:5000
...
  /auth/signup:
    post:
      description: Registers a new user
      consumes:
      - multipart/form-data
      produces:
      - application/json
      parameters:
      - in: formData
        name: email,
        description: Email,
        type: string
      - in: formData
        name: user_password,
        description: Password,
        type: string  
      - in: formData
        name: password2,
        description: Password,
        type: string  
      - in: formData
        name: username,
        description: Username,
        type: string 
      - in: formData
        name: user_address
        description: User Location,
        type: string  
      - in: formData
        name: phone_number
        description: User Contact,
        type: string 
      tags:
        - Authentication
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                user_password:
                  type: string
                user_address:
                  type: string
                phone_number:
                  type: string
                password2:
                  type: string
                email:
                  type: string
                username:
                  type: string
            examples:
              '0':
                value: |-
                  {
                          "username": "Kikonyogo Sadic",
                          "email":"sadic123@gmail.com",
                         "phone_number": "0751547655",
                         "user_address": "Gayaza",
                         "user_password":"#golden@",
                         "password2":"#golden@"
                      
                  }
      responses:
        '200':
          description: Auto generated using Swagger Inspector
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                  message:
                    type: string
                  phone number:
                    type: string
                  passowrd:
                    type: string
                  email:
                    type: string
              examples:
                '0':
                  value: |
                    {
                      "email": "sadic123@gmail.com", 
                      "message": "new user created", 
                      "name": "Kikonyogo Sadic", 
                      "passowrd": "sha256$tykh87GkE7jNLT1w$793ad3304e6792b3de51b12706d43d3bc749a0c8f0ca6d8732f4d12fa0bdc067", 
                      "phone number": "0751547655"
                    }
      servers:
        - url: http://127.0.0.1:5000
    servers:
      - url: http://127.0.0.1:5000
...
components:
...
  securitySchemes:
    oauth2:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: http://yourauthurl.com
          scopes:
            scope_name: Enter your scopes here